db_dive > all getters converted to api calls
